+++ randovania/gui/game_details/base_connection_details_tab.py
@@ -3,16 +3,16 @@ import collections
 from PySide6 import QtWidgets
 
 from randovania.game_description.game_patches import GamePatches
+from randovania.game_description.world.world_list import WorldList
 from randovania.games.game import RandovaniaGame
 from randovania.gui.game_details.game_details_tab import GameDetailsTab
 from randovania.interface_common.players_configuration import PlayersConfiguration
 from randovania.layout import filtered_database
 from randovania.layout.base.base_configuration import BaseConfiguration
 from randovania.lib.dict_lib import iterate_key_sorted
-from randovania.patching.prime import elevators
 
 
-class TeleporterDetailsTab(GameDetailsTab):
+class BaseConnectionDetailsTab(GameDetailsTab):
     def __init__(self, parent: QtWidgets.QWidget, game: RandovaniaGame):
         super().__init__(parent, game)
         self.tree_widget = QtWidgets.QTreeWidget(parent)
@@ -21,9 +21,14 @@ class TeleporterDetailsTab(GameDetailsTab):
         return self.tree_widget
 
     def tab_title(self) -> str:
-        if self.game_enum == RandovaniaGame.METROID_PRIME_CORRUPTION:
-            return "Teleporters"
-        return "Elevators"
+        raise NotImplementedError()
+
+    def _fill_per_world_connections(self,
+                                    per_world: dict[str, dict[str, str]],
+                                    world_list: WorldList,
+                                    patches: GamePatches,
+                                    ):
+        raise NotImplementedError()
 
     def update_content(self, configuration: BaseConfiguration, all_patches: dict[int, GamePatches],
                        players: PlayersConfiguration):
@@ -35,14 +40,7 @@ class TeleporterDetailsTab(GameDetailsTab):
         patches = all_patches[players.player_index]
 
         per_world: dict[str, dict[str, str]] = collections.defaultdict(dict)
-
-        for source, destination_loc in patches.all_elevator_connections():
-            source_world = world_list.world_by_area_location(source.identifier.area_identifier)
-            source_name = elevators.get_elevator_or_area_name(self.game_enum, world_list,
-                                                              source.identifier.area_identifier, True)
-
-            per_world[source_world.name][source_name] = elevators.get_elevator_or_area_name(self.game_enum, world_list,
-                                                                                            destination_loc, True)
+        self._fill_per_world_connections(per_world, world_list, patches)
 
         for world_name, world_contents in iterate_key_sorted(per_world):
             world_item = QtWidgets.QTreeWidgetItem(self.tree_widget)